langchain/requirements.txt
langchain
langchain_community
langchain-core
langchain-text-splitters

langchain_openai
langchain-google-genai

langchain-chroma

google-generativeai

python-dotenv
pypdf
unstructured

InstructorEmbedding
torch
sentence_transformers

faiss-cpu
yfinance
python-weather
duckduckgo-search
ddgs
wikipedia
